+++ changelog.md
@@ -2,6 +2,8 @@
 
 All notable changes to this project will be documented in this file.
 
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
@@ -9,6 +11,19 @@ All notable changes to this project will be documented in this file.
 
 - refactor class name to [BEM](http://getbem.com/naming/)
 - move style to external file
+- update emit name ```hide``` to ```hiddenModal```
+- edit default slot name to body
+- improvement methods component
+
+### Added
+
+- add new emit for finish hidden modal setTimeout ```finishHidden```
+- add header, body and footer slot for more control 
+
+### Removed
+
+- remove default slot
+- remove props zeroSpace and showCloseButton
 
 ### Fixed